+++ test/source/platform/store/contact-store.ts
@@ -1,0 +1,11 @@
+/* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
+
+import { Contact } from '../../core/pgp-key.js';
+
+export class ContactStore {
+
+  public static dbContactGet = async (db: void, emailOrLongid: string[]): Promise<(Contact | undefined)[]> => {
+    return [];
+  }
+
+}
